[APM] moved the apm index_pattern.json back to the core kibana side of things so that it doesn't break the oss only build by requireing a non-existent module.
