+++ src/legacy/core_plugins/kibana/server/tutorials/apm/index_pattern.json
